lecture 138

+++ src/obstacle.rs
@@ -44,11 +44,11 @@ impl Obstacle {
         self.clear_top(self.x_top + OBSTACLE_WIDTH as Coord, self.speed);
         self.clear_bottom(self.x_btm + OBSTACLE_WIDTH as Coord, self.speed);
 
-        if self.x_top <= VIEW_BIGIN - OBSTACLE_WIDTH {
+        if self.x_top <= VIEW_BIGIN - OBSTACLE_WIDTH as i32 {
             self.clear_top(VIEW_BIGIN, OBSTACLE_WIDTH);
         }
 
-        if self.x_btm <= VIEW_BIGIN - OBSTACLE_WIDTH {
+        if self.x_btm <= VIEW_BIGIN - OBSTACLE_WIDTH as i32 {
             self.clear_bottom(VIEW_BIGIN, OBSTACLE_WIDTH);
         }
     }
@@ -79,12 +79,12 @@ impl Obstacle {
         self.draw();
         self.clear();
 
-        if self.x_top <= VIEW_BIGIN - OBSTACLE_WIDTH {
+        if self.x_top <= VIEW_BIGIN - OBSTACLE_WIDTH as i32 {
             self.x_top = VIEW_END;
             self.already_scored = false;
         }
 
-        if self.x_btm <= VIEW_BIGIN - OBSTACLE_WIDTH {
+        if self.x_btm <= VIEW_BIGIN - OBSTACLE_WIDTH as i32 {
             self.x_btm = VIEW_END;
         }
     }
